moved css to assets folder
