+++ client/src/components/SelectQuiz.js
@@ -1,5 +1,5 @@
-import "./css_reset.css"
-import "./styles.css";
+import "./assets/css_reset.css"
+import "./assets/styles.css";
 import React from "react";
 import Button from '@material-ui/core/Button';
 
@@ -16,6 +16,9 @@ function SelectQuiz () {
         </h3>
         <p className="space"><Button type="submit" variant="outlined" color="primary">Random Quiz</Button></p>
         <p className="space"><Button type="submit" variant="outlined" color="primary">Pre-Made Quiz</Button></p>
+        <div className="space"></div>
+        Quiz Id: <input type="text" placeholder="Quiz Id"></input>
+        <div className="space"></div>
         </div>
     )
 }
